% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cw_puma2000_czone}
\alias{cw_puma2000_czone}
\title{PUMA to Commuting Zone Crosswalk (2000 vintage)}
\format{
A data frame with columns:
\describe{
  \item{puma}{6-character PUMA code}
  \item{cz90}{Commuting Zone code}
  \item{afactor}{Allocation factor}
}
}
\source{
Autor, David and David Dorn (2013).
  "The Growth of Low Skill Service Jobs and the Polarization of the U.S. Labor Market."
  *American Economic Review*, 103(5): 1553-1597.
  Crosswalk data available at \url{https://www.ddorn.net/data.htm}.
}
\usage{
cw_puma2000_czone
}
\description{
A dataset containing crosswalk weights from PUMA 2000 to CZ90
}
\keyword{datasets}
