% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cw_puma2020_czone}
\alias{cw_puma2020_czone}
\title{PUMA to Commuting Zone Crosswalk (2020 vintage)}
\format{
A data frame with columns:
\describe{
  \item{puma}{6-character PUMA code}
  \item{cz90}{Commuting Zone code}
  \item{afactor}{Allocation factor}
}
#' @source Constructed by the author, following the
  methodology described in Dorn (2009).
  See Dorn’s original commuting zone crosswalks at \url{https://www.ddorn.net/data.htm}.
}
\usage{
cw_puma2020_czone
}
\description{
A dataset containing crosswalk weights from PUMA 2020 to CZ90
}
\keyword{datasets}
